+++ src/code/util/mod.rs
@@ -20,7 +20,6 @@ mod _use; // _use!
 
 mod asserts; // assertion macros
 mod capture; // capture_[first|last|tail]!
-mod cfg_if; // cfg_if!
 mod cfor; // cfor!
 mod ident; // ident_const_index!
 mod impl_trait; // impl_trait!
@@ -41,7 +40,7 @@ reexports::items! { // structural access: _mods, _internals, _all, _always
 
     mod _mods {
         pub use super::{
-            asserts::_all::*, capture::*, cfg_if::*, cfor::*, ident::*, impl_trait::*, maybe::*,
+            asserts::_all::*, capture::*, cfor::*, ident::*, impl_trait::*, maybe::*,
             methods::*, reexports::*,
         };
         #[cfg(_bit··)]